chore: update dependencies and versions in lock files
Update the versions of several dependencies in the pubspec.lock files 
to ensure compatibility and access to the latest features. The 
dependencies updated include `material_color_utilities`, `meta`, 
`collection`, `get_it`, and others. The version of the example 
project is also incremented to reflect these changes. This improves 
the overall stability and functionality of the project.